// models/_project.ts
/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export const projectModel = {
  languages: {
    default: {
      codename: 'default',
      name: 'Default project language',
    },
  },
  contentTypes: {
    _component__cta: {
      codename: '_component__cta',
      name: '[Component] CTA',
      elements: {
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        button_text: {
          codename: 'button_text',
          name: 'Button text',
        },
        subheadline: {
          codename: 'subheadline',
          name: 'Subheadline',
        },
        button_link: {
          codename: 'button_link',
          name: 'Button link',
        },
      },
    },
    _component__hero: {
      codename: '_component__hero',
      name: '[Component] Hero',
      elements: {
        image: {
          codename: 'image',
          name: 'Image',
        },
        cta_link: {
          codename: 'cta_link',
          name: 'CTA link',
        },
        cta_text: {
          codename: 'cta_text',
          name: 'CTA text',
        },
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        subheadline: {
          codename: 'subheadline',
          name: 'Subheadline',
        },
      },
    },
    _component__text_with_image: {
      codename: '_component__text_with_image',
      name: '[Component] Text with image',
      elements: {
        headline: {
          codename: 'headline',
          name: 'Headline',
        },
        image: {
          codename: 'image',
          name: 'Image',
        },
        text: {
          codename: 'text',
          name: 'Text',
        },
      },
    },
    landing_page: {
      codename: 'landing_page',
      name: 'Landing page',
      elements: {
        content: {
          codename: 'content',
          name: 'Content',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
        url_slug: {
          codename: 'url_slug',
          name: 'URL slug',
        },
      },
    },
  },
  taxonomies: {},
};